pyproject.toml
[project]
name = "mini-order-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt==3.2.2",
    "fastapi>=0.129.0",
    "passlib[bcrypt]==1.7.4",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.22",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.41.0",
]
